@@ -3,11 +3,11 @@ description: A code execution editor which supports multiple languages.
 
 publish_to: 'none'
 
-version: 1.0.0+1
+version: 0.2.0
 
 environment:
   sdk: ">=2.17.0 <3.0.0"
-  flutter: ^3.0.0
+  flutter: ^3.3.0
 
 dependencies:
   flutter:
@@ -35,6 +35,7 @@ dependencies:
   bitsdojo_window: ^0.1.2
   flutter_launcher_icons: ^0.9.3
   connectivity_plus: ^2.3.6+1
+  package_info_plus: ^1.4.2
 
 dev_dependencies:
   flutter_test:
