@@ -1,14 +1,20 @@
 import 'package:dio/dio.dart';
 import 'package:hive_flutter/hive_flutter.dart';
+import 'package:package_info_plus/package_info_plus.dart';
 
 late final Dio dio;
-late Box box;
+late final Box box;
+late final PackageInfo packageInfo;
 
 void initDioClient() {
   dio = Dio();
 }
 
-void initDatabase() async {
+Future<void> initDatabase() async {
   await Hive.initFlutter();
   box = await Hive.openBox('code');
+}
+
+Future<void> initPackageInfo() async {
+  packageInfo = await PackageInfo.fromPlatform();
 }
